awooapi-core/src/main/java/net/fuxle/awooapi/utilities/ISO8601.java
package net.fuxle.awooapi.utilities;

import java.text.ParseException;
import java.text.SimpleDateFormat;
import java.time.LocalDate;
import java.time.LocalDateTime;
import java.time.LocalTime;
import java.time.ZonedDateTime;
import java.time.format.DateTimeFormatter;
import java.util.Calendar;
import java.util.Date;
import java.util.GregorianCalendar;

/**
 * Helper class for handling a most common subset of ISO 8601 strings
 * (in the following format: "2008-03-01T13:00:00+01:00"). It supports
 * parsing the "Z" timezone, but many other less-used features are
 * missing.
 */
public final class ISO8601 {
	/**
	 * Transform Calendar to ISO 8601 string.
	 */
	public static String fromCalendar(final Calendar calendar) {
		Date date = calendar.getTime();
		String formatted = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssZ")
			.format(date);
		return formatted.substring(0, 22) + ":" + formatted.substring(22);
	}

	/**
	 * Get current date and time formatted as ISO 8601 string.
	 */
	public static String now() {
		return fromCalendar(GregorianCalendar.getInstance());
	}

	/**
	 * Transform ISO 8601 string to Calendar.
	 */
	public static Calendar toCalendar(final String iso8601string)
		throws ParseException {
		Calendar calendar = GregorianCalendar.getInstance();
		String s = iso8601string.replace("Z", "+00:00");
		try {
			s = s.substring(0, 22) + s.substring(23);  // to get rid of the ":"
		} catch (IndexOutOfBoundsException e) {
			throw new ParseException("Invalid length", 0);
		}
		Date date = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssZ").parse(s);
		calendar.setTime(date);
		return calendar;
	}

	/**
	 * Transform LocalDate to ISO 8601 string.
	 */
	public static String fromLocalDate(final LocalDate localDate) {
		return localDate.format(DateTimeFormatter.ISO_LOCAL_DATE);
	}

	/**
	 * Transform ISO 8601 string to LocalDate.
	 */
	public static LocalDate toLocalDate(final String iso8601string) {
		return LocalDate.parse(iso8601string, DateTimeFormatter.ISO_LOCAL_DATE);
	}

	/**
	 * Transform LocalDateTime to ISO 8601 string.
	 */
	public static String fromLocalDateTime(final LocalDateTime localDateTime) {
		return localDateTime.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME);
	}

	/**
	 * Transform ISO 8601 string to LocalDateTime.
	 */
	public static LocalDateTime toLocalDateTime(final String iso8601string) {
		if (!iso8601string.contains("T")) {
			throw new IllegalArgumentException("Input must contain both date and time parts separated by 'T'");
		}
		return LocalDateTime.parse(iso8601string, DateTimeFormatter.ISO_LOCAL_DATE_TIME);
	}


	/**
	 * Transform LocalTime to ISO 8601 string.
	 */
	public static String fromLocalTime(final LocalTime localTime) {
		return localTime.format(DateTimeFormatter.ISO_LOCAL_TIME);
	}

	/**
	 * Transform ISO 8601 string to LocalTime.
	 */
	public static LocalTime toLocalTime(final String iso8601string) {
		return LocalTime.parse(iso8601string, DateTimeFormatter.ISO_LOCAL_TIME);
	}

	/**
	 * Transform ISO 8601 string with timezone to LocalDate.
	 */
	public static ZonedDateTime toZonedDateTime(final String iso8601string) {
		return ZonedDateTime.parse(iso8601string, DateTimeFormatter.ISO_DATE_TIME);
	}


}
